update evm router
