+++ packages/bitcore-node/src/modules/arbitrum/api/arb-routes.ts
@@ -1,4 +1,4 @@
-import { EVMRouter } from '../../ethereum/api/evmRouter';
+import { EVMRouter } from '../../../providers/chain-state/evm/api/routes';
 import { ARB } from './csp';
 
 class ARBRouter extends EVMRouter {
